TeamLLM1/requirements.txt
# Core ML (버전 고정)
torch==2.4.1
numpy==1.26.4
transformers==4.39.3
sentence-transformers==2.2.2
safetensors==0.4.2
huggingface-hub==0.23.0
faiss-cpu

# .env
python-dotenv

# IO / utils
pandas
tqdm
gdown
# API & Web
fastapi
uvicorn[standard]
pydantic

# OpenAI API
openai

# LangGraph
langgraph
langchain-core>=0.2
langsmith>=0.1.95